[level 2] Title: k진수에서 소수 개수 구하기, Time: 16.26 ms, Memory: 80.2 MB -BaekjoonHub

+++ 프로그래머스/2/92335. k진수에서 소수 개수 구하기/k진수에서 소수 개수 구하기.java
@@ -24,8 +24,8 @@ class Solution {
     }
     
     public boolean isPrime(long n) {
-        long max = (long) Math.sqrt(n);
-        for (long i=2;i<=max;i++) {
+        int max = (int) Math.sqrt(n);
+        for (int i=2;i<=max;i++) {
             if (n % i == 0) {
                 return false;
             }
